Directional Exploration: First pass XML parsing & file IO
Still Buggy, lots of low-hanging fruit.

Better way to handle tags is to create a stack of tags and have the top
one always the active. We push a tag on opening, pop a tag on closing.
Right now there is some shitty logic scraped together to get it working,
but it doesn't handle nested tags at all, hence the stack scheme.

Second big thing on the list is the various string functions, if we want
to keep them and avoid using the "string.h" class (which I would not
mind at all) those need to be cleaned up and refined.